requirements.txt
django
django-import-export
crispy-bootstrap5
django-extensions
django-import-export
django-registration-redux
razorpay
django-versatileimagefield